+++ examples/taro-react/CHANGELOG.md
@@ -18,14 +18,14 @@
 - Updated dependencies
   - @captcha-pro/taro-react@2.3.0
 
-## 2.0.3
+## 2.2.0
 
 ### Patch Changes
 
 - Updated dependencies
   - @captcha-pro/taro-react@2.2.0
 
-## 2.0.2
+## 2.1.0
 
 ### Patch Changes